Naming improvements
- rename `apply.InstanceApplier` to `reconciler.Reconciler`
- use separate files inside the package
- move interactive functions to dedicated `reconciler.InteractiveReconciler` type
